+++ ui/frontend/src/main/src/app/shared/query-gateway/query-gateway.module.ts
@@ -1,5 +1,7 @@
+import { HttpClient } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { EpicProvider } from '../redux-utils/provider';
 import { QueryGatewayService } from './query-gateway.service';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { BrowserModule } from '@angular/platform-browser';
@@ -7,10 +9,10 @@ import { CovalentDialogsModule } from '@covalent/core';
 import { AppMaterialModule } from '../app-material-module';
 import { QueryProgressComponent } from './query-progress/query-progress.component';
 import { QueryErrorComponent } from './query-error/query-error.component';
-import { QueryEpicProvider } from './query.redux';
+import { fallsQueryAngefordert} from './query.redux';
 
 @NgModule({
-    providers: [QueryGatewayService, QueryEpicProvider],
+    providers: [QueryGatewayService, EpicProvider(fallsQueryAngefordert, [HttpClient])],
     imports: [
         CommonModule,
         BrowserModule,
